+++ rocksdb/src/test/java/io/telicent/smart/cache/storage/rocksdb/External.java
@@ -10,6 +10,7 @@ import org.rocksdb.RocksDBException;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 public class External extends AbstractRocksDBStorage {