+++ CHANGELOG.md
@@ -7,6 +7,7 @@
         - `close()` now actively persists any created counters in case they have been used in a non-transactional
           manner.
         - Added `getDefaultHandle()` method for obtaining the default column family handle.
+        - Added `dropColumnFamily()` method for dropping column families.
         - Improved various aspects of error messages and error handling.
         - Added some basic diagnostic logging
     - Transaction improvements:
@@ -18,12 +19,15 @@
           committed/closed
         - Added `forEach()` method to `TransactionContext` to allow full iteration and processing of a column family,
           intended primarily for one-off expensive operations like data migration
-        - Added `iterator()` method to `TransacationContext` to allow full iteration control over a column family
+        - Added `iterator()` method to `TransacationContext` to allow full iterator access to a column family
     - Label Store improvements:
         - Column Family handle names are now protected constants meaning storage derived from the `RocksDBLabelsStore`
           can access and manipulate those column handles where needed
+- Build improvements:
+    - Upgraded RocksDB to 10.5.1
 
 # 0.9.1
+
 - Build improvements:
     - Addressing CVE-2026-1605 (Jetty)